move template tests to where they belong
