+++ packages/templates/typescript/package.json
@@ -10,9 +10,11 @@
     "test": "codegen-handlebars-templates-scripts test"
   },
   "devDependencies": {
+    "@types/common-tags": "1.4.0",
     "codegen-handlebars-templates-scripts": "0.8.21",
-    "graphql-codegen-core": "0.8.21",
-    "graphql-codegen-compiler": "0.8.21"
+    "common-tags": "1.7.2",
+    "graphql-codegen-compiler": "0.8.21",
+    "graphql-codegen-core": "0.8.21"
   },
   "main": "./dist/index.js",
   "typings": "dist/index.d.ts",
